fix(location): address template review feedback
