+++ app/src/main/java/com/noobexon/xposedfakelocation/manager/ui/settings/SettingsViewModel.kt
@@ -95,6 +95,33 @@ class SettingsViewModel(application: Application) : AndroidViewModel(application
         }
     }
 
+    private class StringPreference(
+        initialValue: String,
+        private val flow: Flow<String>,
+        private val saveOperation: suspend (String) -> Unit,
+        private val viewModelScope: kotlinx.coroutines.CoroutineScope
+    ) {
+        private val _state = MutableStateFlow(initialValue)
+        val state: StateFlow<String> = _state.asStateFlow()
+
+        init {
+            viewModelScope.launch {
+                flow.collect { _state.value = it }
+            }
+        }
+
+        fun setValue(value: String) {
+            _state.value = value
+            viewModelScope.launch {
+                try {
+                    saveOperation(value)
+                } catch (e: Exception) {
+                    // Add error handling if needed
+                }
+            }
+        }
+    }
+
     // Preferences for Accuracy
     private val _useAccuracyPreference = BooleanPreference(
         DEFAULT_USE_ACCURACY,
@@ -154,16 +181,15 @@ class SettingsViewModel(application: Application) : AndroidViewModel(application
     )
     val useGpsNoise: StateFlow<Boolean> = _useGpsNoisePreference.state
 
-    private val _gpsNoiseLevel = MutableStateFlow(GpsNoiseLevel.NORMAL)
-    val gpsNoiseLevel: StateFlow<GpsNoiseLevel> = _gpsNoiseLevel.asStateFlow()
-
-    init {
-        viewModelScope.launch {
-            preferencesRepository.getGpsNoiseLevelFlow().collect { value ->
-                _gpsNoiseLevel.value = parseGpsNoiseLevel(value)
-            }
-        }
-    }
+    private val _gpsNoiseLevelPreference = StringPreference(
+        DEFAULT_GPS_NOISE_LEVEL,
+        preferencesRepository.getGpsNoiseLevelFlow(),
+        preferencesRepository::saveGpsNoiseLevel,
+        viewModelScope
+    )
+    val gpsNoiseLevel: StateFlow<GpsNoiseLevel> = _gpsNoiseLevelPreference.state
+        .map(GpsNoiseLevel::fromPreferenceValue)
+        .stateIn(viewModelScope, SharingStarted.WhileSubscribed(5000), GpsNoiseLevel.NORMAL)
 
     // Preferences for Vertical Accuracy
     private val _useVerticalAccuracyPreference = BooleanPreference(
@@ -285,12 +311,7 @@ class SettingsViewModel(application: Application) : AndroidViewModel(application
     fun setUseRandomize(value: Boolean) = _useRandomizePreference.setValue(value)
     fun setRandomizeRadius(value: Double) = _randomizeRadiusPreference.setValue(value)
     fun setUseGpsNoise(value: Boolean) = _useGpsNoisePreference.setValue(value)
-    fun setGpsNoiseLevel(value: GpsNoiseLevel) {
-        _gpsNoiseLevel.value = value
-        viewModelScope.launch {
-            preferencesRepository.saveGpsNoiseLevel(value.name)
-        }
-    }
+    fun setGpsNoiseLevel(value: GpsNoiseLevel) = _gpsNoiseLevelPreference.setValue(value.name)
     fun setUseVerticalAccuracy(value: Boolean) = _useVerticalAccuracyPreference.setValue(value)
     fun setVerticalAccuracy(value: Float) = _verticalAccuracyPreference.setValue(value)
     fun setUseMeanSeaLevel(value: Boolean) = _useMeanSeaLevelPreference.setValue(value)
@@ -304,8 +325,4 @@ class SettingsViewModel(application: Application) : AndroidViewModel(application
     fun setHideFakeLocationToast(value: Boolean) = _hideFakeLocationToastPreference.setValue(value)
     fun setUseInAppTargetApps(value: Boolean) = _useInAppTargetAppsPreference.setValue(value)
     fun setEnableBroadcastControl(value: Boolean) = _enableBroadcastControlPreference.setValue(value)
-
-    private fun parseGpsNoiseLevel(value: String): GpsNoiseLevel {
-        return runCatching { GpsNoiseLevel.valueOf(value) }.getOrDefault(GpsNoiseLevel.NORMAL)
-    }
 }
