+++ app/src/main/java/com/noobexon/xposedfakelocation/manager/ui/targetapps/TargetAppsViewModel.kt
@@ -210,7 +210,7 @@ class TargetAppsViewModel(application: Application) : AndroidViewModel(applicati
             useSpeedAccuracy = preferencesRepository.getUseSpeedAccuracy(),
             speedAccuracy = preferencesRepository.getSpeedAccuracy(),
             useGpsNoise = preferencesRepository.getUseGpsNoise(),
-            gpsNoiseLevel = parseGpsNoiseLevel(preferencesRepository.getGpsNoiseLevel())
+            gpsNoiseLevel = GpsNoiseLevel.fromPreferenceValue(preferencesRepository.getGpsNoiseLevel())
         )
     }
 
@@ -223,9 +223,5 @@ class TargetAppsViewModel(application: Application) : AndroidViewModel(applicati
             .map { it.packageName }
             .toSet()
         preferencesRepository.saveTargetApps(selectedPackages)
-    }
-
-    private fun parseGpsNoiseLevel(value: String): GpsNoiseLevel {
-        return runCatching { GpsNoiseLevel.valueOf(value) }.getOrDefault(GpsNoiseLevel.NORMAL)
     }
 }
