+++ app/src/main/java/com/noobexon/xposedfakelocation/manager/ui/templates/TemplatesViewModel.kt
@@ -61,7 +61,7 @@ class TemplatesViewModel(application: Application) : AndroidViewModel(applicatio
                     useSpeedAccuracy = preferencesRepository.getUseSpeedAccuracy(),
                     speedAccuracy = preferencesRepository.getSpeedAccuracy(),
                     useGpsNoise = preferencesRepository.getUseGpsNoise(),
-                    gpsNoiseLevel = parseGpsNoiseLevel(preferencesRepository.getGpsNoiseLevel())
+                    gpsNoiseLevel = GpsNoiseLevel.fromPreferenceValue(preferencesRepository.getGpsNoiseLevel())
                 )
             )
         }
@@ -86,9 +86,5 @@ class TemplatesViewModel(application: Application) : AndroidViewModel(applicatio
         viewModelScope.launch {
             preferencesRepository.removeLocationTemplate(template.id)
         }
-    }
-
-    private fun parseGpsNoiseLevel(value: String): GpsNoiseLevel {
-        return runCatching { GpsNoiseLevel.valueOf(value) }.getOrDefault(GpsNoiseLevel.NORMAL)
     }
 }
